requirements.txt
g2p_en==2.1.0
inflect==5.3.0
jieba==0.42.1
numpy
onnxruntime==1.10.0
paddlepaddle==2.2.1
pypinyin==0.44.0
SoundFile==0.10.3.post1
